scripts/train_calibration.py
import pandas as pd
import numpy as np
import yaml
import joblib
from pathlib import Path
from polymarket_quant.calibration.models import SegmentedCalibrator
from polymarket_quant.evaluation.metrics import calibration_diagnostics

def run_training():
    # Load data (Assuming processed Parquet from ingestion)
    data_path = Path("data/processed/markets_normalized_latest.parquet")
    if not data_path.exists():
        print("No data found. Run ingestion first.")
        return

    df = pd.read_parquet(data_path)
    
    # 1. LEAKAGE PREVENTING SPLIT
    # Use markets resolved before 2025 for training
    cutoff = pd.Timestamp("2025-01-01", tz="UTC")
    train_df = df[df['resolution_date'] < cutoff].copy()
    test_df = df[df['resolution_date'] >= cutoff].copy()

    print(f"Training on {len(train_df)} markets. Testing on {len(test_df)}.")

    # 2. FIT CALIBRATOR
    calibrator = SegmentedCalibrator()
    # Mocking price/outcome for this skeleton; in reality, join with trade data
    # outcomes = train_df['resolved_as_yes'].values
    # prices = train_df['last_price'].values
    # ...
    
    # 3. EVALUATE
    # diag = calibration_diagnostics(test_probs, test_outcomes)
    # print(f"Brier Score: {diag['brier']:.4f}")
    
    # 4. SAVE ARTIFACTS
    artifact_dir = Path("artifacts/calibration")
    artifact_dir.mkdir(parents=True, exist_ok=True)
    joblib.dump(calibrator, artifact_dir / "segmented_calibrator.pkl")

if __name__ == "__main__":
    run_training()