pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polymarket_quant"
version = "0.1.0"
description = "Production-grade quantitative research framework for Polymarket"
authors = [{ name = "Quant Researcher" }]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.2.0",
    "scipy>=1.10.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pyarrow>=15.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "isort>=5.0",
    "mypy>=1.0",
]
